requirements.txt
pandas
scikit-learn
nltk
imapclient
pyzmail36